+++ ChangeLog.md
@@ -1,5 +1,6 @@
 ## Changes between ifm3d 0.1.0 and 0.2.0
 
+* Added software trigger support to O3X
 * Optimization to `ifm3d` cmd line tool when passed either `--help` or
   `version`. It will no longer try to connect to the device first, which makes
   this much more responsive and convenient for when no h/w is plugged in.